feat : 게시글 추가 기능 완료

+++ step17-board-backend/src/main/java/com/spring/dto/BoardDTO.java
@@ -25,7 +25,7 @@ public class BoardDTO {
   private int bcount;
   private int blike;
   private int bhate;
-  private int mid;
+  private Long mid;
   private String nickname;
 
 }
